# docs/swagger.yaml
basePath: /
definitions:
...
  main.listModel:
    properties:
      action:
        example: all|backup|export|import|ip|jail|limit|log|path|port|prio|release|state|template|type
        format: string
        type: string
      options:
        example: -d|-j|-p|-u|-x
        format: string
        type: string
    type: object
...
  main.templateModel:
    properties:
      action:
        example: --convert
        format: string
        type: string
      options:
        example: -a|-x
        format: string
        type: string
      target:
        example: jail_target
        format: string
        type: string
      template:
        example: template/name
        format: string
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    API interface to FreeBSD bastille
    Observation: Do not use console, htop and top commands with API. Only for UI Interface.
  license:
    name: BSD-3-Clause
    url: https://opensource.org/license/bsd-3-clause
  termsOfService: http://swagger.io/terms/
  title: Bastille-Web
  version: "1.0"
paths:
...
  /list:
    post:
      consumes:
      - application/json
      description: List jails, ports, releases, templates, logs, limits, exports and
        imports and much more managed by bastille.
      parameters:
      - description: list
        in: body
        name: list
        required: true
        schema:
          $ref: '#/definitions/main.listModel'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: list command
      tags:
      - list
...
  /template:
    post:
      consumes:
      - application/json
      description: Apply file templates to targeted container(s).
      parameters:
      - description: template
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/main.templateModel'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: template command
      tags:
      - template
...
swagger: "2.0"